Moved tests into proper folder Removed unused dlls max pain level is based on TOU. fix spell points on character load. fix paint icon after a fill operation. fix non-image shadow drawing (non-selectable hexes)

+++ src/main/java/ostrowski/combat/common/CombatMap.java
@@ -2336,7 +2336,7 @@ public class CombatMap extends SerializableObject implements Enums, IMonitorable
       }
       Node nodeAlpha = namedNodeMap.getNamedItem("backgroundImageAlpha");
       if (nodeAlpha != null) {
-         _backgroundImageAlpha = Byte.parseByte(nodeAlpha.getNodeValue());
+         _backgroundImageAlpha = Integer.parseInt(nodeAlpha.getNodeValue());
       }
 
       _bgImageUpperLeft = new Point(0,0);
